Solving Errors

+++ MVP/EcareWebApp/src/main/java/com/ecare/repositories/ExchangeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ecare.models.Exchange;
 
 @Repository
-public interface ExchangeRepository extends CrudRepository<Exchange, Double> {
+public interface ExchangeRepository extends CrudRepository<Exchange, Long> {
 
 }
