+++ MVP/EcareWebApp/src/main/java/com/ecare/repositories/WalletOneRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.ecare.models.WalletOne;
 
 @Repository
-public interface WalletOneRepository extends CrudRepository<WalletOne, Double>{
+public interface WalletOneRepository extends CrudRepository<WalletOne, Long>{
 	List<WalletOne>findAll();
 }
